rewrite EstimateDropdown w/ ComboBox
